Fixed tests and issues
