+++ src/main/java/com/synopsys/integration/alert/web/audit/AuditEntryController.java
@@ -26,6 +26,7 @@ package com.synopsys.integration.alert.web.audit;
 import java.util.Optional;
 import java.util.UUID;
 
+import org.apache.commons.lang3.BooleanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -51,8 +52,7 @@ public class AuditEntryController extends BaseController {
     private ResponseFactory responseFactory;
 
     @Autowired
-    public AuditEntryController(final AuditEntryActions auditEntryActions, final ContentConverter contentConverter,
-        final ResponseFactory responseFactory) {
+    public AuditEntryController(final AuditEntryActions auditEntryActions, final ContentConverter contentConverter, final ResponseFactory responseFactory) {
         this.auditEntryActions = auditEntryActions;
         this.contentConverter = contentConverter;
         this.responseFactory = responseFactory;
@@ -62,7 +62,7 @@ public class AuditEntryController extends BaseController {
     public AlertPagedModel<AuditEntryModel> get(@RequestParam(value = "pageNumber", required = false) final Integer pageNumber, @RequestParam(value = "pageSize", required = false) final Integer pageSize,
         @RequestParam(value = "searchTerm", required = false) final String searchTerm, @RequestParam(value = "sortField", required = false) final String sortField,
         @RequestParam(value = "sortOrder", required = false) final String sortOrder, @RequestParam(value = "onlyShowSentNotifications", required = false) final Boolean onlyShowSentNotifications) {
-        return auditEntryActions.get(pageNumber, pageSize, searchTerm, sortField, sortOrder, onlyShowSentNotifications);
+        return auditEntryActions.get(pageNumber, pageSize, searchTerm, sortField, sortOrder, BooleanUtils.toBoolean(onlyShowSentNotifications));
     }
 
     @GetMapping(value = "/{id}")
